requirements.txt
Django==1.9.1
requests==2.9.1
beautifulsoup4==4.4.1
celery==3.1.20
django-celery==3.1.17
factory_boy==2.6.0
redis==2.10.5
pytz==2015.7
